finish tidalsim script extraction
